+++ src/app/modules/slormancer/constants/content/data/default-configs.ts
@@ -164,6 +164,7 @@ export const DEFAULT_CONFIG: CharacterConfig = {
     highest_slorm_temple_floor: 0,
     show_temple_keeper_as_totem: false,
     show_elder_inner_fire_damage: false,
+    chilled_enemy_nearby: 0,
 }
 
 export const COMBAT_CONFIG: CharacterConfig = {
@@ -328,4 +329,5 @@ export const COMBAT_CONFIG: CharacterConfig = {
     highest_slorm_temple_floor: 0,
     show_temple_keeper_as_totem: false,
     show_elder_inner_fire_damage: true,
+    chilled_enemy_nearby: 0,
 }