+++ src/app/modules/slormancer/model/character-config.ts
@@ -162,5 +162,6 @@ export interface CharacterConfig {
     highest_slorm_temple_floor: number;
     show_temple_keeper_as_totem: boolean;
     show_elder_inner_fire_damage: boolean;
+    chilled_enemy_nearby: number;
 }