+++ src/app/modules/slormancer/model/content/data/data-mechanic.ts
@@ -1,7 +1,9 @@
 import { AbstractEffectValue } from '../effect-value';
 import { SkillGenre } from '../enum/skill-genre';
+import { SkillElement } from '../skill-element';
 
 export interface DataMechanic {
     values: Array<AbstractEffectValue>;
     genres?: Array<SkillGenre>;
+    element?: SkillElement;
 }