+++ src/app/modules/slormancer/model/content/mechanic.ts
@@ -1,6 +1,7 @@
 import { AbstractEffectValue } from './effect-value';
 import { MechanicType } from './enum/mechanic-type';
 import { SkillGenre } from './enum/skill-genre';
+import { SkillElement } from './skill-element';
 
 export interface Mechanic {
     type: MechanicType;
@@ -8,6 +9,7 @@ export interface Mechanic {
     icon: string;
     description: string;
     genres: Array<SkillGenre>;
+    element: SkillElement;
 
     template: string;
     values: Array<AbstractEffectValue>;